Refactor LanguageServerExtensionService to be a generic implementation
