+++ core/esmf-aspect-meta-model-java/pom.xml
@@ -79,6 +79,10 @@
          <groupId>org.jspecify</groupId>
          <artifactId>jspecify</artifactId>
       </dependency>
+      <dependency>
+         <groupId>net.harawata</groupId>
+         <artifactId>appdirs</artifactId>
+      </dependency>
 
       <!-- Test dependencies -->
       <dependency>
