+++ core/esmf-turtle-language-server/src/main/java/org/eclipse/esmf/turtle/languageserver/lsp/ResolutionStrategyService.java
@@ -21,6 +21,7 @@ import org.eclipse.esmf.aspectmodel.resolver.fs.FlatModelsRoot;
 import org.eclipse.esmf.aspectmodel.resolver.fs.StructuredModelsRoot;
 import org.eclipse.esmf.aspectmodel.resolver.github.GitHubStrategy;
 import org.eclipse.esmf.aspectmodel.resolver.github.GithubModelSourceConfig;
+import org.eclipse.esmf.aspectmodel.service.ExtensionService;
 import org.eclipse.esmf.turtle.languageserver.aspect.MetaModelStrategy;
 import org.eclipse.esmf.turtle.languageserver.aspect.navigation.GithubMaterializingStrategy;
 import org.eclipse.esmf.turtle.languageserver.lsp.config.GithubResolutionConfiguration;
@@ -61,7 +62,7 @@ public class ResolutionStrategyService {
       strategies.add( structuredStrategy );
       strategies.add( flatStrategy );
       strategies.add( metaModelStrategy );
-      strategies.addAll( LanguageServerExtensionService.getResolutionStrategies() );
+      strategies.addAll( ExtensionService.getResolutionStrategies() );
       for ( final GithubModelSourceConfig repository : githubConfiguration.repositories() ) {
          strategies.add( githubStrategyCache.computeIfAbsent( repository,
                config -> new GithubMaterializingStrategy( new GitHubStrategy( config ) ) ) );
